pyproject.toml
[project]
name = "ai-quant-company"
version = "0.1.0"
description = "Multi-Agent 量化公司仿真与自动研究系统"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AI Quant Team"}
]
keywords = ["quantitative", "trading", "multi-agent", "AI", "backtest"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web 框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    
    # 数据库
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    
    # 数据处理
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    
    # LLM 客户端
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    
    # 加密货币数据
    "ccxt>=4.2.0",
    "python-binance>=1.0.19",
    
    # 报告生成
    "jinja2>=3.1.3",
    "markdown>=3.5.0",
    "weasyprint>=61.0",
    
    # 工具库
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.26.0",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    
    # 可视化
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
